Leave clippy with nothing to say
Nine warnings that had been piling up: a redundant cast, a string copied
twice, two repeat().take(), a manual range check and two unused bindings.
All of them mechanical, none of them mine from today, and every one an
equivalent rewrite.

The one about trabajar() taking eight arguments gets an allow with the
reason written down instead: the last five are the counters this thread
shares with whoever reads it from outside, and boxing them to please a
lint would be a wrapper you have to open at every use.

+++ src-tauri/src/exporter.rs
@@ -1493,7 +1493,7 @@ mod medir {
         let indices: Vec<usize> = (0..cuantos).collect();
         let delays: Vec<u32> = indices.iter().map(|_| 33).collect();
         let mut loader = |i: usize| -> Result<image::RgbaImage> {
-            let d = (i as i64 % 20) as i64 - 10;
+            let d = (i as i64 % 20) - 10;
             let mut movida = image::RgbaImage::new(origen.width(), origen.height());
             image::imageops::replace(&mut movida, &origen, d, d / 2);
             Ok(movida)
